+++ cli/helpers/server_runner.go
@@ -12,8 +12,9 @@ type HooksServerRunConfig struct {
 	WunderGraphDirAbs string
 	ServerScriptFile  string
 	Env               []string
-	LogLevel          string
+	Log               string
 	Production        bool
+	DebugBindAddress  string
 	Debug             bool
 	Output            *scriptrunner.OutputConfig
 }
@@ -28,15 +29,15 @@ func NewHooksServerRunner(log *zap.Logger, cfg *HooksServerRunConfig) *scriptrun
 		hooksEnv = append(hooksEnv, "NODE_ENV=production")
 	}
 
-	// Align the log level
-	if cfg.LogLevel != "" {
-		hooksEnv = append(hooksEnv, fmt.Sprintf("WG_LOG_LEVEL=%s", cfg.LogLevel))
+	// Align the log configuration
+	if cfg.Log != "" {
+		hooksEnv = append(hooksEnv, fmt.Sprintf("WG_LOG=%s", cfg.Log))
 	}
 
 	var scriptArgs []string
 
 	if cfg.Debug {
-		scriptArgs = append(scriptArgs, "--inspect")
+		scriptArgs = append(scriptArgs, "--inspect="+cfg.DebugBindAddress)
 	}
 	scriptArgs = append(scriptArgs, cfg.ServerScriptFile)