+++ packages/sdk/src/codegen/templates/typescript/jsonschema.template.ts
@@ -3,6 +3,13 @@ export const template = `
 // @ts-ignore: no-types available		
 import type {JSONSchema7} from "json-schema";
 
+// @ts-ignore: module unavailable
+declare module 'json-schema' {
+	export interface JSONSchema7 {
+		'x-graphql-enum-name'?: string;
+	}
+}
+
 export interface Queries {
     {{#each queries}}
     "{{name}}": {