+++ packages/sdk/src/integrations/datasources/index.ts
@@ -1,2 +1,13 @@
 export { type GraphQLDatasourceOptions, graphql } from './graphql';
-export { defineDatasource } from './define-datasource';
+export { type MongoDBDatasourceOptions, mongodb } from './mongodb';
+export { type MySQLDatasourceOptions, mysql } from './mysql';
+export { type NatsKVDatasourceOptions, natsKV } from './natskv';
+export { type OpenAPIDatasourceOptions, openapi } from './openapi';
+export { type PlanetScaleDatasourceOptions, planetscale } from './planetscale';
+export { type PostgreSQLDatasourceOptions, postgresql } from './postgresql';
+export { type PrismaDatasourceOptions, prisma } from './prisma';
+export { type SQLiteDatasourceOptions, sqlite } from './sqlite';
+export { type SQLServerDatasourceOptions, sqlserver } from './sqlserver';
+export { type FederationDatasourceOptions, federation } from './federation';
+
+export { defineDatasource } from '../define-datasource';