+++ packages/sdk/src/logger/index.ts
@@ -35,32 +35,39 @@ const resolvePinoLogLevel = (level: string): PinoLogLevel => {
 			return PinoLogLevel.Info;
 		case LogLevel.DEBUG:
 			return PinoLogLevel.Debug;
+		case LogLevel.TRACE:
+			return PinoLogLevel.Trace;
 		default:
 			throw new Error(`Unknown log level: ${wgLogLevel}`);
 	}
 };
 
 const initLogger = (destination: DestinationStream): pino.Logger => {
-	const logLevel =
-		process.env.WG_DEBUG_MODE === 'true'
-			? PinoLogLevel.Debug
-			: process.env.WG_CLI_LOG_LEVEL
-			? resolvePinoLogLevel(process.env.WG_CLI_LOG_LEVEL)
-			: PinoLogLevel.Info;
-
+	const logLevel = resolvePinoLogLevel(process.env.WG_LOG ?? 'info');
 	let options: pino.LoggerOptions = {
 		...pinoOptions(),
 		level: logLevel,
 	};
 
-	return pino(options);
+	return pino(options, destination);
 };
 
 const logger = initLogger(process.stdout);
 
+const cloudBindings: Record<string, any> = {};
+if (process.env.WG_CLOUD_ENVIRONMENT_ID) {
+	cloudBindings.environmentID = process.env.WG_CLOUD_ENVIRONMENT_ID;
+}
+if (process.env.WG_CLOUD_PROJECT_ID) {
+	cloudBindings.projectID = process.env.WG_CLOUD_PROJECT_ID;
+}
+if (process.env.WG_CLOUD_DEPLOYMENT_ID) {
+	cloudBindings.deploymentID = process.env.WG_CLOUD_DEPLOYMENT_ID;
+}
+
 export const Logger = logger.child({ component: '@wundergraph/sdk' });
 export const ServerLogger = logger.child(
-	{ component: '@wundergraph/server' },
+	{ component: '@wundergraph/server', ...cloudBindings },
 	{ level: process.env.WG_DEBUG_MODE === 'true' ? PinoLogLevel.Debug : PinoLogLevel.Info }
 );